pyproject.toml
[tool.poetry]
name = "gryphon-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Michael D. M. Dryden <mk.dryden@utoronto.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {version = "^20.6", extras = ["job-queue"]}
typer = {version = "^0.9.0", extras = ["all"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

